+++ relm4-examples/examples/non_blocking_async.rs
@@ -1,12 +1,15 @@
 use gtk::prelude::{BoxExt, ButtonExt, GtkWindowExt, OrientableExt};
-use relm4::{send, AppUpdate, Model, RelmApp, Sender, WidgetPlus, Widgets};
+use relm4::{
+    send, AppUpdate, Components, MessageHandler, Model, RelmApp, RelmMsgHandler, Sender,
+    WidgetPlus, Widgets,
+};
 
-use tokio::runtime::Builder;
+use tokio::runtime::{Builder, Runtime};
 use tokio::sync::mpsc::{channel, Sender as TokioSender};
 
-struct AppModel {
-    counter: u8,
-    async_handler: TokioSender<(AsyncHandlerMsg, Sender<AppMsg>)>,
+struct AsyncHandler {
+    _rt: Runtime,
+    sender: TokioSender<AsyncHandlerMsg>,
 }
 
 #[derive(Debug)]
@@ -15,20 +18,87 @@ enum AsyncHandlerMsg {
     DecrementRequest,
 }
 
+impl MessageHandler<AppModel> for AsyncHandler {
+    type Msg = AsyncHandlerMsg;
+    type Sender = TokioSender<AsyncHandlerMsg>;
+
+    fn init(_parent_model: &AppModel, parent_sender: Sender<AppMsg>) -> Self {
+        let (sender, mut rx) = channel::<AsyncHandlerMsg>(10);
+
+        let rt = Builder::new_multi_thread()
+            .worker_threads(8)
+            .enable_time()
+            .build()
+            .unwrap();
+
+        rt.spawn(async move {
+            while let Some(msg) = rx.recv().await {
+                let parent_sender = parent_sender.clone();
+                tokio::spawn(async move {
+                    tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+                    match msg {
+                        AsyncHandlerMsg::IncrementRequest => {
+                            send!(parent_sender, AppMsg::Increment);
+                        }
+                        AsyncHandlerMsg::DecrementRequest => {
+                            send!(parent_sender, AppMsg::Decrement);
+                        }
+                    }
+                });
+            }
+        });
+
+        AsyncHandler { _rt: rt, sender }
+    }
+
+    fn send(&self, msg: Self::Msg) {
+        self.sender.blocking_send(msg).unwrap();
+    }
+
+    fn sender(&self) -> Self::Sender {
+        self.sender.clone()
+    }
+}
+
+struct AppModel {
+    counter: u8,
+}
+
 #[derive(Debug)]
 enum AppMsg {
     Increment,
     Decrement,
 }
 
+struct AppComponents {
+    async_handler: RelmMsgHandler<AsyncHandler, AppModel>,
+}
+
+impl Components<AppModel> for AppComponents {
+    fn init_components(
+        parent_model: &AppModel,
+        _parent_widget: &AppWidgets,
+        parent_sender: Sender<AppMsg>,
+    ) -> Self {
+        AppComponents {
+            async_handler: RelmMsgHandler::new(parent_model, parent_sender),
+        }
+    }
+}
+
 impl Model for AppModel {
     type Msg = AppMsg;
     type Widgets = AppWidgets;
-    type Components = ();
+    type Components = AppComponents;
 }
 
 impl AppUpdate for AppModel {
-    fn update(&mut self, msg: AppMsg, _components: &(), _sender: Sender<AppMsg>) -> bool {
+    fn update(
+        &mut self,
+        msg: AppMsg,
+        _components: &AppComponents,
+        _sender: Sender<AppMsg>,
+    ) -> bool {
         match msg {
             AppMsg::Increment => {
                 self.counter = self.counter.wrapping_add(1);
@@ -55,13 +125,13 @@ impl Widgets<AppModel, ()> for AppWidgets {
 
                 append = &gtk::Button {
                     set_label: "Increment",
-                    connect_clicked(sender, async_sender) => move |_| {
-                        async_sender.blocking_send((AsyncHandlerMsg::IncrementRequest, sender.clone())).unwrap();
+                    connect_clicked[sender = components.async_handler.sender()] => move |_| {
+                        sender.blocking_send(AsyncHandlerMsg::IncrementRequest).expect("Receiver dropped");
                     },
                 },
                 append = &gtk::Button::with_label("Decrement") {
-                    connect_clicked(sender, async_sender) => move |_| {
-                        async_sender.blocking_send((AsyncHandlerMsg::DecrementRequest, sender.clone())).unwrap();
+                    connect_clicked[sender = components.async_handler.sender()] => move |_| {
+                        sender.blocking_send(AsyncHandlerMsg::DecrementRequest).expect("Receiver dropped");
                     },
                 },
                 append = &gtk::Label {
@@ -71,41 +141,10 @@ impl Widgets<AppModel, ()> for AppWidgets {
             },
         }
     }
-
-    fn pre_init() {
-        let async_sender = &model.async_handler;
-    }
 }
 
 fn main() {
-    let (rx, mut tx) = channel::<(AsyncHandlerMsg, Sender<AppMsg>)>(10);
-
-    let rt = Builder::new_multi_thread()
-        .worker_threads(8)
-        .enable_time()
-        .build()
-        .unwrap();
-
-    rt.spawn(async move {
-        while let Some((msg, sender)) = tx.recv().await {
-            tokio::spawn(async move {
-                tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-                match msg {
-                    AsyncHandlerMsg::IncrementRequest => {
-                        send!(sender, AppMsg::Increment);
-                    }
-                    AsyncHandlerMsg::DecrementRequest => {
-                        send!(sender, AppMsg::Decrement);
-                    }
-                }
-            });
-        }
-    });
-
-    let model = AppModel {
-        counter: 0,
-        async_handler: rx,
-    };
+    let model = AppModel { counter: 0 };
     let app = RelmApp::new(model);
     app.run();
 }